Hard assertions completed
